Test the program update cronjob importing the event program
The existing program update tests only check that the job logged completion, so nothing caught the import dropping a batch or clearing a start time the lottery had closed. This drives the real job on its own schedule alongside the assignment one, which is why the file is now named for the pair.
