+++ server/CLAUDE.md
@@ -234,7 +234,10 @@ When writing `signedToStartTime` for sign-ups, store the program item's own `sta
 - **Logging:** winston logger in `server/src/utils/logger.ts` — human-readable in dev, JSON in production; error-level logs are also forwarded to Sentry (`winston-transport-sentry-node`). Console level is `info` by default; the `LOG_LEVEL` env var overrides it (`debug` enables debug logs, and the Playwright scripts set `warn` to hide per-request logs). Logging is mocked off in tests unless explicitly enabled in config. Log errors by passing the `Error` directly — `logger.error(new Error("context"))`, or wrap an underlying error as the cause — `logger.error(new Error("context", { cause: err }))`. **Do not use the old `logger.error("%s", new Error(...))` idiom**: the `%s` is unnecessary (the logger surfaces the Error to Sentry and renders the stack/cause itself) and stringifying the Error into the message produced a doubled `Error: Error:` prefix in Sentry.
 - **Error handling:** the final Express error middleware (in `utils/server.ts`) logs and responds 500; a JSON body-parse guard handles malformed bodies as 400; `@sentry/node`'s express error handler captures exceptions. Prefer returning `Result` errors from services over throwing.
 - **Sentry:** initialized in `utils/instrument.ts`; DSN/sample rate from `config.sentry()`. Under ESM, importing `instrument.ts` from `index.ts` is **not enough** to auto-instrument Express — by the time `Sentry.init()` runs, Express is already linked and can't be wrapped (you'd see `[Sentry] express is not instrumented`). The `start` / `start:dev` / `start:test` scripts therefore preload it with Node's `--import=./src/utils/instrument.ts` flag (via `tsx`), which registers the OpenTelemetry import hooks before the app module graph loads. The `release` is read from `process.env.APP_VERSION` (set to the git SHA in the runner image, matching the release `deploy.yml` creates in `konsti-backend-{env}`); it is the Sentry release name only and is **not** part of the settings response, which reports `appBuildTime` — the image's build time, and the sole deploy identifier the client sees; it's `undefined` locally. Unlike the frontend maps/release — which upload during the image **build** — the backend release is created **after** a successful `deploy` (the `backend-sentry-release` job, `needs: deploy`), so a failed deploy creates no backend release.
-- **Cron jobs:** scheduled in `utils/cron.ts` using `croner` (e.g. auto-updating program items from Kompassi and auto-running assignment), toggled per-environment via server config; guards prevent a stale server instance from running a job. The assignment job passes `assignmentTime: null`, so **which** start time it lotteries comes from `getDynamicStartTime` (`getTimeNow()` + `directSignupPhaseStart`) and therefore honours the mocked clock, while **when** it fires comes from croner and the real one. The two clocks are what `cron.test.ts` and `cronBatchedLottery.test.ts` split between them: the first calls the job handler directly under `vi.setSystemTime` to cover the instance check, the lock and job registration; the second registers the real job on a one-second schedule and must **not** freeze the clock, because a frozen date leaves croner waiting out the delay to the next real tick. It seeds the live event's own fleamarket batch from the Kompassi dump (`useLocalProgramFile`) so a parent hour that stopped lining up with its sub-sessions fails a test rather than an event.
+- **Cron jobs:** scheduled in `utils/cron.ts` using `croner` (e.g. auto-updating program items from Kompassi and auto-running assignment), toggled per-environment via server config; guards prevent a stale server instance from running a job. The assignment job passes `assignmentTime: null`, so **which** start time it lotteries comes from `getDynamicStartTime` (`getTimeNow()` + `directSignupPhaseStart`) and therefore honours the mocked clock, while **when** it fires comes from croner and the real one. The two clocks are what `cron.test.ts` and `cronEventProgram.test.ts` split between them. `cron.test.ts` calls the job handlers directly under `vi.setSystemTime`, covering the instance check, the lock and job registration. `cronEventProgram.test.ts` registers the real jobs on a one-second schedule and must **not** freeze the clock, because a frozen date leaves croner waiting out the delay to the next real tick; it runs both jobs against the live event config and the event's own Kompassi dump (`useLocalProgramFile`), so a parent hour or a slug that stopped lining up with its sub-sessions fails a test rather than an event. Each of its tests registers only the job it drives, since the other ticking every second alongside would re-import the program mid-lottery.
+
+  The two jobs' last-run times are not the same kind of marker, which matters when waiting on one: `assignmentLastRun` is written after the run, but `programUpdateLastRun` **is** the update's concurrency guard (a tick within 30 seconds of the last is refused as one still in flight) and so is written before the import starts. Wait on the update's completion log, not its last-run time.
+
 - **Email notifications:** queued through `utils/notificationQueue.ts` (a `fastq` queue) and sent via `nodemailer`.
 - **Security & static serving:** `helmet` for headers; the SPA in `server/front/` is served with `express-static-gzip` (brotli/gzip). Cache policy: files served from the bundle's `assets/` directory get a one-year immutable `Cache-Control`, everything else (`index.html`, `robots.txt`, the SPA fallback) gets `no-cache`. Clients therefore revalidate the HTML on every load but keep old chunks cached across deploys, so a stale page still finds its matching chunks instead of fetching deleted files. The rule keys on the directory rather than the shape of the filename because a name alone is ambiguous - an ordinary hyphenated name is indistinguishable from a hashed one - and the directory is matched relative to the served root, so an `assets` directory somewhere above it can't widen the rule.
 - **Mongoose conventions:** global plugins in `db/mongoosePlugins.ts` (lean getters/virtuals via `mongoose-lean-getters`/`mongoose-lean-virtuals`, `toJSON` transforms that strip `_id`/`__v`). Mongoose applies global plugins at model-compile time, so **every file calling `mongoose.model()` imports `server/db/mongoosePlugins` for its side effect as its first import** — a model compiled before that module runs silently loses the plugins, and DB reads/writes then fail in ways that look unrelated. Add that import to any new schema file rather than relying on some other module having pulled it in first. Connection lifecycle in `db/mongodb.ts`. Dates are `Date`, handled with date-fns (see [shared/CLAUDE.md](../shared/CLAUDE.md)).
